requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
supabase>=2.0.0
python-dotenv>=1.0.0
fastmcp>=0.1.0
slack-sdk>=3.27.0
anthropic>=0.30.0
pydantic>=2.0.0
